+++ src/pages/collage/CollagePageLegacy.js
@@ -34,58 +34,14 @@ import { useNavigate, useLocation } from 'react-router-dom';
 import { LoadingButton } from "@mui/lab";
 import { useTheme } from "@mui/material/styles";
 import { API } from 'aws-amplify';
-import BasePage from "./BasePage";
-import { UserContext } from "../UserContext";
-import { useSubscribeDialog } from "../contexts/useSubscribeDialog";
-
-// Utility function to hash username for localStorage (same as in CollagePage)
-const hashString = (str) => {
-  let hash = 0;
-  if (str.length === 0) return hash;
-  for (let i = 0; i < str.length; i += 1) {
-    const char = str.charCodeAt(i);
-    hash = ((hash * 33) - hash) + char;
-    hash = Math.imul(hash, 1); // Convert to 32bit integer
-  }
-  return Math.abs(hash).toString();
-};
-
-// Utility functions for localStorage preference management (same as in CollagePage)
-const getCollagePreferenceKey = (user) => {
-  if (!user?.userDetails?.email) return 'memeSRC-collage-preference-anonymous';
-  const hashedUsername = hashString(user.userDetails.email);
-  return `memeSRC-collage-preference-${hashedUsername}`;
-};
-
-const setCollagePreference = (user, preference) => {
-  const key = getCollagePreferenceKey(user);
-  localStorage.setItem(key, preference);
-};
-
-// Utility functions for managing banner dismiss timestamp
-const getBannerDismissKey = (user) => {
-  if (!user?.userDetails?.email) return 'memeSRC-banner-dismiss-anonymous';
-  const hashedUsername = hashString(user.userDetails.email);
-  return `memeSRC-banner-dismiss-${hashedUsername}`;
-};
-
-const isBannerDismissed = (user) => {
-  const key = getBannerDismissKey(user);
-  const dismissedTimestamp = localStorage.getItem(key);
-  if (!dismissedTimestamp) return false;
-  
-  const now = new Date().getTime();
-  const dismissedTime = parseInt(dismissedTimestamp, 10);
-  const sevenDaysInMs = 7 * 24 * 60 * 60 * 1000; // 7 days in milliseconds
-  
-  return (now - dismissedTime) < sevenDaysInMs;
-};
-
-const setBannerDismissed = (user) => {
-  const key = getBannerDismissKey(user);
-  const now = new Date().getTime().toString();
-  localStorage.setItem(key, now);
-};
+import BasePage from "../BasePage";
+import { UserContext } from "../../UserContext";
+import { useSubscribeDialog } from "../../contexts/useSubscribeDialog";
+import {
+  setCollagePreference,
+  isBannerDismissed,
+  setBannerDismissed
+} from "../../utils/collagePreferences";
 
 // Component for subtle chip to try new version when banner is dismissed
 function SubtleSwitchChip({ onTryNewVersion }) {
