+++ src/pages/collage/index.js
@@ -1,0 +1,2 @@
+export { default as CollagePage } from './CollagePage';
+export { default as CollagePageLegacy } from './CollagePageLegacy';
